reorganised Config.cpp & Blocks.cpp and divided them into several Config_x.cpp & Blocks_x.cpp, for code fragmentation. Also renamed testsBlocks_parsing.cpp to Blocks_parsing.test.cpp, to follow conventions.
